log: remove private global logger
It was originally put in place in PBnJ to be used by log.GetPackaged
which would return a copy of the global/original/root logger. That api
didn't live very long and was changed to logger.Package instead, in case
we want to keep context from a parent logger in the copy/child logger.