log: skip 1 caller in Fatal path
zap will add `caller:"dir/file.go:lineno"` to the log message. Since
`Fatal` calls `Error` for code re-use we need to account for this new
uninteresting intermediate frame. `Init` has already setup skip=1 to
skip Debug/Info/Error calls correctly, but that means when we call
`Fatal` zap logs it instead of the *reall* caller! We add one more to
the skip so that the caller of `Fatal` is what ends up in the log
message.