Update rollbar
I had to remove the WithMessage test from TestPkgErrorsCompat because the new
rollbar logic + test isn't quite 1:1 to the old setup. rError was digging
through all the errors until it found a stackTracer. The new rollbar code does
the digging outside of `StackTracer` (the replacement for rErrors' Stack).
WithMessage thus fails the test because StackTracer does not do any digging.

So its no longer in unit tests but it was added to examplelog so we can still
verify nothing breaks, albeitly via a human.