rollbar: Remove call to SetServerRoot
The rollbar ui has stopped correctly parsing/displaying our project
frames within the stacktrace. This is *not* due to the recent update
to the rollbar-go package, I was able to see this on master at the time
too. I suspect this was introduced when being used in a go module project.