logr/packetlogger: logr, zapr deps updated to v1.2.2
Note: This is a breaking change, hence versioned under /v2

With this change, returning *PacketLogr as a logr.Logger isn't
possible, since logr.Logger is of type struct and it does not provide
the methods that old logr.Logger interface would declare.

The NewPacketLogr() is updated to return type *PacketLogr
which embeds the logr.Logger struct type, it inherits various methods
of the logr.Logger, although methods expecting a logr.Logger would need to
reference the *PacketLogr.Logger struct field.

https://github.com/go-logr/logr/blob/v1.2.2/logr.go#L230
https://github.com/go-logr/logr/blob/v0.4.0/logr.go#L148

The go-logr/logr v0.4.0 was an BETA API,
https://github.com/go-logr/logr/blob/v0.4.0/logr.go#L20